sched/tune: beautify schedtune_tasks_update()
The schedtune_tasks_update() function is used to update the accouting
of tasks on a certain boost group. It support only add/remove operations
and provide proper trace events as well as CPU boost update if
required.

This patch does not change the funtional bits but contributs to cleanin
up a bit the code to make it more simple and readable from call sites.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>