drm/msm/sde: fix plane stage setup in sde crtc
SDE crtc handles the plane stage based on mixer
loop. That adds an issue when two planes are staged
on same z_order. It overwrites the second plane
configuration with first one. That leads to underrun.

This cleans up the issues in and around plane stage
setup to avoid overwrite or unnecessary locking. It
fixes the issue for source split cases.

Change-Id: I2e784ab4fbb06b8156bef3a8d53a5ae8d69b2fd3
Signed-off-by: Dhaval Patel <pdhaval@codeaurora.org>