cpufreq: cpu-boost: Only boost certain input events
The events to boost are based on HTC changes to the kernel although
the implementation of how to do is different in this commit.

The touch driver generates a lot of additional events and in some
very basic instrumentation I saw them more then 100ms after
touching the screen.  Filtering these out seems like a good idea.

Change-Id: Icc7ee16ca03659716e80a72590bb2ade9e63824e
Signed-off-by: Eliminater74 <eliminater74@gmail.com>