ANDROID: fiq_debugger: Fix minor bug in code
We fix a typo in the code which had us comparing a pointer instead
of the value which was being pointed to.  This turns out to be
a relatively benign bug, as we'd incorrectly pass in the empty
string instead of NULL to the function, and the function can handle
both.  But we fix it so the code is clearly doing what we intend.

Signed-off-by: Greg Kaiser <gkaiser@google.com>
Change-Id: Ib059819775a3bebca357d4ce684be779853156e3
Signed-off-by: Tab Fitts <tfitts1212@gmail.com>