pme: defconfig: update for rcu updates
Latest changes in RCU should now improve the conditions for enabling
RCU boosting without much power costs.

Priority: 6.

This level of PRIO should be the proper spot on latency sensitive
workloads like a mobile platform with improved RT consumers in
userspace. A value of 2 could also be acceptable for those targets.

NO-CB: All [0-3 for MSM8996].

Processes RCU callbacks in different threads, in this case we allow all
the four CPUs we have available on this target to "share" their load when
high intensity RCU callbacks kick-in, a condition that could lead to RCU
starvation without it (overload of the softirq mechanism).